Add new training datasets for model training: z_training_75.json and z_training_all.json
- Introduced z_training_75.json containing detailed probability distributions for 'yes' and 'no' responses across various features.
- Introduced z_training_all.json with similar structure, providing comprehensive data for model training.
- Both files include target variable percentages and a list of columns for reference.
