+++ Credit_Card_Applications_training_all.json
@@ -1,0 +1,4 @@
+{
+    "yes": {
+        "CustomerID": {
+
